+++ src/ralph/api/auth/oidc.py
@@ -1,8 +1,9 @@
 """OpenID Connect authentication tool for the Ralph API."""
 
+import base64
 import logging
 from functools import lru_cache
-from typing import Dict, Optional, Union, Literal, get_args
+from typing import Dict, Literal, Optional, Union, get_args
 
 import requests
 from fastapi import Depends, HTTPException, status
@@ -27,30 +28,60 @@ logger = logging.getLogger(__name__)
 
 
 class UserInfo(BaseModel):
-    """Pydantic model representing the core of an OpenID Connect UserInfo endpoint response.
+    """Pydantic model representing the UserInfo response of the OIDC IdP.
 
     They are common to both the ID token and the UserInfo endpoint.
-
+    We do not use it for authentication, so may claims are ignored.
     They are polymorphic and may have many attributes not defined in the
     specification. This model ignores all additional fields.
 
     Attributes:
-        iss (str): Issuer Identifier for the Issuer of the response.
         sub (str): Subject Identifier.
-        aud (str or list of str): Audience(s) that this ID Token/UserInfo is intended for.
-        exp (int): Expiration time on or after which the ID Token/UserInfo MUST NOT be
+        scope (str): Scope(s) for resource authorization.
+        target (str): Target for storing the statements (custom claim).
+    """
+
+    sub: str
+    scope: Optional[str] = None
+    target: Optional[str] = None
+
+    model_config = ConfigDict(extra="ignore")
+
+
+class TokenInfo(BaseModel):
+    """Pydantic model representing the Introspection response of the OIDC IdP.
+
+    Based on the RFC 7662 section 2.2 definition of token /introspect response
+    This model does not use all fields defined in the RFC,
+    and we force some optional fields to be present.
+
+    The 'active' field is assumed to be true and is not included in this model.
+
+    Attributes:
+        client_id (str): ID of the client that owns this token
+        username (str): Name of the user referred to by this this token, if any
+        iss (str): Issuer Identifier for the Issuer of the response.
+        sub (str): Subject Identifier, if any
+                  No sub means that this is a purely 'client' token
+        aud (str or list of str): Audience(s) that this ID Token is intended for.
+        exp (int): Expiration time on or after which the ID Token MUST NOT be
                    accepted for processing.
         iat (int): Time at which the JWT was issued.
         scope (str): Scope(s) for resource authorization.
-        target (str): Target for storing the statements.
+        target (str): Target for storing the statements (custom claim).
+
     """
 
+    client_id: str
+    username: Optional[str] = None
+    token_type: Optional[str] = None
     iss: str
-    sub: str
+    sub: Optional[str] = None
     aud: Optional[Union[list[str], str]] = None
     exp: float
     iat: float
     scope: Optional[str] = None
+
     target: Optional[str] = None
 
     model_config = ConfigDict(extra="ignore")
@@ -73,9 +104,12 @@ def discover_provider(base_url: AnyUrl) -> Dict:
             headers={"WWW-Authenticate": "Bearer"},
         ) from exc
 
+
 def get_user_info(provider_config: dict, access_token: str) -> UserInfo:
     """Get the user's info from the IdP using the /userinfo OIDC endpoint."""
-    user_info, is_encoded = get_user_info_data(provider_config["userinfo_endpoint"], access_token)
+    user_info, is_encoded = get_user_info_data(
+        provider_config["userinfo_endpoint"], access_token
+    )
     if not is_encoded:
         # nothing to do
         return UserInfo.model_validate(user_info)
@@ -83,10 +117,13 @@ def get_user_info(provider_config: dict, access_token: str) -> UserInfo:
 
 
 @lru_cache()
-def get_user_info_data(userinfo_endpoint: AnyUrl, access_token: str) -> Union[tuple[dict, Literal[False]], tuple[str, Literal[True]]]:
+def get_user_info_data(
+    userinfo_endpoint: AnyUrl, access_token: str
+) -> Union[tuple[dict, Literal[False]], tuple[str, Literal[True]]]:
     """Get the user's info from the IdP using the /userinfo OIDC endpoint.
 
-    The data may be unencoded (Content-Type: 'application/json', in which case it is a json dictionary
+    The data may be unencoded (Content-Type: 'application/json',
+    in which case it is a json dictionary.
     If it is encoded (Content-Type: 'application/jwt', it is a JWT
     see: https://openid.net/specs/openid-connect-core-1_0.html#UserInfo
 
@@ -137,6 +174,58 @@ def decode_user_info(encoded_user_info: str, provider_config: dict) -> UserInfo:
         ) from exc
 
     return UserInfo.model_validate(decoded_token)
+
+
+def encode_client_secret_basic_token(client_id: str, client_secret: str) -> str:
+    """Encode client id and client secret inside an opaque token.
+
+    Should be sent as Authorization: Basic {token}
+
+    This corresponds to the `client_secret_basic`
+    token endpoint authentication method.
+    """
+    return base64.b64encode(
+        client_id.encode("utf-8") + b":" + client_secret.encode("utf-8")
+    ).decode("utf-8")
+
+
+@lru_cache()
+def get_token_info(
+    introspection_endpoint: AnyUrl, token: str, client_id: str, client_secret: str
+) -> UserInfo:
+    """Get info on given token from the IdP using /introspection OIDC endpoint."""
+    token_info = None
+    try:
+        response = requests.post(
+            f"{introspection_endpoint}",
+            headers={
+                "Authorization": f"Basic {encode_client_secret_basic_token(
+                    client_id=client_id,
+                    client_secret=client_secret
+                )}"
+            },
+            data={
+                "token": f"{token}",
+            },
+            timeout=5,
+        )
+        response.raise_for_status()
+        token_info = response.json()
+    except requests.exceptions.RequestException as exc:
+        logger.error("Unable to get token info: %s", exc)
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Could not validate credentials",
+            headers={"WWW-Authenticate": "Bearer"},
+        ) from exc
+    if not token_info["active"]:
+        logger.error("Inactive or invalid token info.")
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Could not validate credentials",
+            headers={"WWW-Authenticate": "Bearer"},
+        )
+    return TokenInfo.model_validate(token_info)
 
 
 @lru_cache()
@@ -199,11 +288,36 @@ def get_oidc_user(
 
     access_token = auth_header.split(" ")[-1]
     provider_config = discover_provider(settings.RUNSERVER_AUTH_OIDC_ISSUER_URI)
-    user_info = get_user_info(
-        provider_config, access_token=access_token
+
+    token_info = get_token_info(
+        provider_config["introspection_endpoint"],
+        token=access_token,
+        client_id=settings.RUNSERVER_AUTH_OIDC_CLIENT_ID,
+        client_secret=settings.RUNSERVER_AUTH_OIDC_CLIENT_SECRET,
     )
-    return AuthenticatedUser(
-        agent={"openid": f"{user_info.iss}/{user_info.sub}"},
-        scopes=get_user_scopes(user_info.scope),
-        target=user_info.target,
-    )
+    if token_info.sub:
+        # This is a real user, we can retrieve their user info
+        user_info = get_user_info(provider_config, access_token=access_token)
+        if user_info.sub != token_info.sub:
+            logger.error(
+                ("Inconsistent token subject: %s != %s"), user_info.sub, token_info.sub
+            )
+            raise HTTPException(
+                status_code=status.HTTP_401_UNAUTHORIZED,
+                detail="Could not validate credentials",
+                headers={"WWW-Authenticate": "Bearer"},
+            )
+
+        return AuthenticatedUser(
+            agent={"openid": f"{token_info.iss}/user/{user_info.sub}"},
+            scopes=get_user_scopes(user_info.scope),
+            target=user_info.target,
+        )
+    else:
+        # this is an application token, we don't have a user to get
+        # so we use the client_id to indentify it instead
+        return AuthenticatedUser(
+            agent={"openid": f"{token_info.iss}/application/{token_info.client_id}"},
+            scopes=get_user_scopes(token_info.scope),
+            target=token_info.target,
+        )